First pass new light client
