+++ Cargo.toml
@@ -3,7 +3,6 @@ members = [
     "cli",
     "codegen",
     "lightclient",
-    "lightclient2",
     "testing/substrate-runner",
     "testing/test-runtime",
     "testing/integration-tests",
