+++ subxt/src/backend/rpc/mod.rs
@@ -60,6 +60,10 @@ crate::macros::cfg_jsonrpsee! {
     mod jsonrpsee_impl;
 }
 
+crate::macros::cfg_unstable_light_client! {
+    mod lightclient_impl;
+}
+
 crate::macros::cfg_reconnecting_rpc_client! {
     mod reconnecting_jsonrpsee_impl;
     pub use reconnecting_jsonrpsee_ws_client as reconnecting_rpc_client;
